requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycryptodomex==3.9.7
pyjwkest==1.4.2
python-dateutil==2.8.1
python-taiga==1.0.0
requests==2.23.0
six==1.15.0
urllib3==1.25.9
Werkzeug==1.0.1
coverage==5.1
pytest==5.4.2
